[LLK] Address review: fold the alpha sweep into test_sfpu_ema.py, add empty aliases
Two review comments.

nstojicTT, helpers/perf/test_schemas.py: the perf_sfpu_ema entry was the only one
of the 24 in the catalog without an "aliases" key. Added it as an empty dict — it
is genuinely empty at version 1, this is purely so the catalog stays uniform.

skrsmanovicTT, test_sfpu_ema_alpha_sweep.py: it drives the same kernel through the
same TestConfig as the functional test, so it belongs in the same file rather than
a second one. Merged into test_sfpu_ema.py, with the device run factored into
_run_ema_on_device() and shared by both tests; that removes the duplicated
stimuli/tilize/TestConfig block the split had forced. The sweep keeps its nightly
marker, so the default suite still selects 3 tests and deselects the 1000.

Mixing a nightly test with default-suite ones in a single file is the existing
convention here — test_sfpu_reduce.py, test_eltwise_unary_sfpu.py and
test_sfpu_ternary.py all do it.

Behaviour is unchanged: same goldens, same bounds, same LLK_EMA_SWEEP_DUMP hook.

Verified on Wormhole n300: test_sfpu_ema 3 passed, test_perf_header_gate 17
passed, and a sweep slice reproduces the characterisation in the PR body
(alpha=0.962 -> 3.763 half-ULPs = 1.88 ULP, the documented worst case).
